Update bats instructions and move to a README file
The bats tests reference the original, unmaintained, repository
for the bats framework at https://github.com/sstephenson/bats
which is incompatible with the tests in this repository.

Change the instructions to reference the maintained repository
for the bats framework at https://github.com/bats-core/bats-core.

Put these instructions into a single README file, so that if they need
to be changed again it will only have to be done in one place.