Rewrite alpine bats testing script
Keeping both scripts because I plan to merge
them into something better.

Not sure why the script needs a complete rewrite
but here we are.

The new script will work but it won't be elegant.
Elegance will come in (hopefully) the next commit.

Naming the new script `gitwatch-dev` as it is moving towards
something more general than just running bats test;
it provides a container to poke gitwatch in.